google_ocr/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub==1.7.0
google-cloud-vision==1.0.0
